requirements.txt
arabic-reshaper==3.0.0
clipboard==0.0.4
fonttools==4.53.1
numpy==2.0.0
opencv-python==4.10.0.84
packaging==24.1
pillow==10.4.0
pyperclip==1.9.0
pytesseract==0.3.10
python-bidi==0.4.2
six==1.16.0
